// frontend/src/app/modules/core/components/ms-dialog/models/interfaces/dialog-message-config.interface.ts
//   Copyright 2024 Cisco Systems, Inc.
//
//   Licensed under the Apache License, Version 2.0 (the "License");
//   you may not use this file except in compliance with the License.
//   You may obtain a copy of the License at
//
//        http://www.apache.org/licenses/LICENSE-2.0
//
//   Unless required by applicable law or agreed to in writing, software
//   distributed under the License is distributed on an "AS IS" BASIS,
//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//   See the License for the specific language governing permissions and
//   limitations under the License.
//
//   SPDX-License-Identifier: Apache-2.0

import { DialogMessageTypeEnum } from '../enums/dialog-message-type.enum';

export interface DialogMessageConfig {
	message?: string;
	type?: DialogMessageTypeEnum;
}